+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CourseRestController.java
@@ -2,6 +2,9 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.CourseDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.StudentDTO;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.CourseDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.SchoolDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.StudentNotExistsException;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.CourseController;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import org.springframework.web.bind.annotation.*;
@@ -18,37 +21,37 @@ public class CourseRestController {
     }
 
     @PostMapping("/{schoolId}")
-    public Course createCourse(@RequestBody CourseDTO courseDTO, @PathVariable Long schoolId){
+    public Course createCourse(@RequestBody CourseDTO courseDTO, @PathVariable Long schoolId) throws SchoolDoNotExistException {
         return courseController.createCourse(courseDTO, schoolId);
     }
 
     @PutMapping("/{courseId}")
-    public Course updateCourseName(@PathVariable Long courseId, @RequestBody CourseDTO courseDTO){
+    public Course updateCourseName(@PathVariable Long courseId, @RequestBody CourseDTO courseDTO) throws CourseDoNotExistException {
         return courseController.updateCourseName(courseId, courseDTO);
     }
 
     @PutMapping("/{courseId}/addStudent/{studentId}")
-    public Course addStudentToCourse(@PathVariable Long courseId, @PathVariable Long studentId){
+    public Course addStudentToCourse(@PathVariable Long courseId, @PathVariable Long studentId) throws CourseDoNotExistException, StudentNotExistsException {
         return courseController.addStudentToCourse(courseId, studentId);
     }
 
     @PutMapping ("/{courseId}/addStudentByDNI/{studentDNI}")
-    public Course addStudentToCourseByDNI(@PathVariable Long courseId, @PathVariable String studentDNI){
+    public Course addStudentToCourseByDNI(@PathVariable Long courseId, @PathVariable String studentDNI) throws CourseDoNotExistException, StudentNotExistsException {
         return courseController.addStudentToCourseByDNI(courseId, studentDNI);
     }
 
     @PutMapping("/{courseId}/addStudentList")
-    public Course addStudentListToCourse(@PathVariable Long courseId, @RequestBody List<StudentDTO> studentListDTO){
+    public Course addStudentListToCourse(@PathVariable Long courseId, @RequestBody List<StudentDTO> studentListDTO) throws CourseDoNotExistException {
         return courseController.addStudentListToCourse(courseId, studentListDTO);
     }
 
     @PutMapping("/{oldCourseId}/migrateToCourse/{newCourseId}")
-    public Course migrateStudentsToCourse(@PathVariable Long oldCourseId, @PathVariable Long newCourseId){
+    public Course migrateStudentsToCourse(@PathVariable Long oldCourseId, @PathVariable Long newCourseId) throws CourseDoNotExistException {
         return courseController.migrateStudentsToCourse(oldCourseId, newCourseId);
     }
 
     @GetMapping("/{courseId}")
-    public Course getCourseById(@PathVariable Long courseId){
+    public Course getCourseById(@PathVariable Long courseId) throws CourseDoNotExistException {
         return courseController.getCourseById(courseId);
     }