+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/exceptions/CourseDoNotExistException.java
@@ -1,0 +1,7 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions;
+
+public class CourseDoNotExistException extends Exception{
+    public CourseDoNotExistException(String message) {
+        super(message);
+    }
+}