+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/CourseController.java
@@ -2,6 +2,9 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.CourseDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.StudentDTO;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.CourseDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.SchoolDoNotExistException;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions.StudentNotExistsException;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
@@ -26,30 +29,43 @@ public class CourseController {
         this.studentService = studentService;
     }
 
-    public Course createCourse(CourseDTO courseDTO, Long schoolId) {
+    public Course createCourse(CourseDTO courseDTO, Long schoolId) throws SchoolDoNotExistException {
         School school = schoolService.getSchoolById(schoolId);
+        if(school==null){
+            throw new SchoolDoNotExistException("School with id "+ schoolId +" does not exist");
+        }
         return courseService.createCourse(courseDTO, school);
 
     }
 
-    public Course updateCourseName(Long courseId, CourseDTO courseDTO) {
+    public Course updateCourseName(Long courseId, CourseDTO courseDTO) throws CourseDoNotExistException {
         Course course = courseService.getCourseById(courseId);
+        if(course==null){
+            throw new CourseDoNotExistException("Course with id "+ courseId +" does not exist");
+        }
         course.setCourseName(courseDTO.getCourseName());
         return courseService.saveCourse(course);
     }
 
-    public Course addStudentToCourse(Long courseId, Long studentId) {
+    public Course addStudentToCourse(Long courseId, Long studentId) throws CourseDoNotExistException, StudentNotExistsException {
         Course course = courseService.getCourseById(courseId);
         Student student = studentService.getStudentById(studentId);
-        if(student!=null && course!=null){
-            student.setCourse(course);
-            course.addStudent(student);
-            courseService.saveCourse(course);
+        if(student==null){
+            throw new StudentNotExistsException("Student with id "+ studentId +" does not exist");
         }
+        if(course==null){
+            throw new CourseDoNotExistException("Course with id "+ courseId +" does not exist");
+        }
+        student.setCourse(course);
+        course.addStudent(student);
         return courseService.saveCourse(course);
+
     }
 
-    public Course getCourseById(Long courseId) {
+    public Course getCourseById(Long courseId) throws CourseDoNotExistException {
+        if(courseService.getCourseById(courseId)==null){
+            throw new CourseDoNotExistException("Course with id "+ courseId +" does not exist");
+        }
         return courseService.getCourseById(courseId);
     }
 
@@ -63,8 +79,12 @@ public class CourseController {
         courseService.deleteCourse(course);
     }
 
-    public Course addStudentListToCourse(Long courseId, List<StudentDTO> studentListDTO) {
+    public Course addStudentListToCourse(Long courseId, List<StudentDTO> studentListDTO) throws CourseDoNotExistException {
         Course course = courseService.getCourseById(courseId);
+        if(course==null){
+            throw new CourseDoNotExistException("Course with id "+ courseId +" does not exist");
+        }
+
         for (StudentDTO studentDTO : studentListDTO) {
             if(checkStudentExists(studentDTO)){
                 System.err.println("Student with DNI:"+ studentDTO.getStudentDNI() +"already exists");
@@ -87,33 +107,35 @@ public class CourseController {
         return studentService.getStudentByDNI(studentDTO.getStudentDNI())!=null;
     }
 
-    public Course addStudentToCourseByDNI(Long courseId, String studentDNI) {
+    public Course addStudentToCourseByDNI(Long courseId, String studentDNI) throws CourseDoNotExistException, StudentNotExistsException {
         Student student = studentService.getStudentByDNI(studentDNI);
-        if(student!=null){
-            Course course = courseService.getCourseById(courseId);//TODO: CHECK IF COURSE EXISTS
+        Course course = courseService.getCourseById(courseId);
+        if(student==null){
+            throw new StudentNotExistsException("Student with DNI "+ studentDNI +" does not exist");
+        }
+        if(course==null){
+            throw new CourseDoNotExistException("Course with id "+ courseId +" does not exist");
+        }
+
             student.setCourse(course);
             course.addStudent(student);
             return courseService.saveCourse(course);
-        }else{
-            System.err.println("Student with DNI:"+ studentDNI +" not found");
-            return null;
-        }
     }
 
-    public Course migrateStudentsToCourse(Long oldCourseId, Long newCourseId) {
+    public Course migrateStudentsToCourse(Long oldCourseId, Long newCourseId) throws CourseDoNotExistException {
         Course cOld = courseService.getCourseById(oldCourseId);
         Course cNew = courseService.getCourseById(newCourseId);
-        if(cOld!=null && cNew!=null){
-            List<Student> students = cOld.getStudents();
-            for(Student student: students){
-                student.setCourse(cNew);
-                cNew.addStudent(student);
-            }
-            cOld.cleanCourse();
-            courseService.saveCourse(cNew);
-            return cNew;
+        if(cOld==null || cNew==null){
+            throw new CourseDoNotExistException("Course with id "+ oldCourseId +" or "+ newCourseId +" does not exist");
         }
-        System.err.println("Course not found");
-        return null;
+
+        List<Student> students = cOld.getStudents();
+        for(Student student: students){
+            student.setCourse(cNew);
+            cNew.addStudent(student);
+        }
+        cOld.cleanCourse();
+        courseService.saveCourse(cNew);
+        return cNew;
     }
 }