+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/SchoolService.java
@@ -19,7 +19,9 @@ public class SchoolService {
         return schoolRepository.save(school);
     }
 
-    public School getSchoolById(Long schoolId) {return schoolRepository.findById(schoolId).get();}
+    public School getSchoolById(Long schoolId) {
+        return schoolRepository.findById(schoolId).orElse(null);
+    }
     public void deleteSchool(Long schoolId) {schoolRepository.deleteById(schoolId);}
 
     public Optional<Object> getSchoolByName(String schoolName) {